Backend Server/requirements.txt
paho-mqtt
flask
mysql-connector